Added better error handling around DS conversion, and failed DS conversions will properly rollback if the conversion fails.

+++ data/store/convert.go
@@ -20,7 +20,10 @@ import (
 
 // convert converts from the old datastore type (cznic/kv) to the new one (boltdb)
 func convert(filename string) error {
-	kvFile := filepath.Join(filepath.Dir(filename), "."+filepath.Base(filename)+".converted")
+	kvFile := convertedFile(filename)
+	if fileExists(kvFile) {
+		return fmt.Errorf("Old convered datastore file already exists, Datastore %s is already converted!", kvFile)
+	}
 	err := os.Rename(filename, kvFile)
 	if err != nil {
 		return err
@@ -33,6 +36,7 @@ func convert(filename string) error {
 		VerifyDbAfterClose:  true})
 
 	if err != nil {
+		revertFiles(filename)
 		return err
 	}
 	defer kvStore.Close()
@@ -41,18 +45,21 @@ func convert(filename string) error {
 
 	boltStore.StrictMode = true
 	if err != nil {
+		revertFiles(filename)
 		return err
 	}
 	defer boltStore.Close()
 
 	enum, err := kvStore.SeekFirst()
 	if err != nil {
+		revertFiles(filename)
 		return err
 	}
 
 	err = boltStore.Update(func(tx *bolt.Tx) error {
 		bucket, err := tx.CreateBucketIfNotExists([]byte(filename))
 		if err != nil {
+			revertFiles(filename)
 			return err
 		}
 		for {
@@ -61,23 +68,27 @@ func convert(filename string) error {
 				break
 			}
 			if err != nil {
+				revertFiles(filename)
 				return err
 			}
 
 			err = bucket.Put(k, v)
 			if err != nil {
+				revertFiles(filename)
 				return err
 			}
 		}
 		return nil
 	})
 	if err != nil {
+		revertFiles(filename)
 		return err
 	}
 
 	err = boltStore.View(func(tx *bolt.Tx) error {
 		kvIter, err := kvStore.SeekFirst()
 		if err != nil {
+			revertFiles(filename)
 			return err
 		}
 
@@ -90,22 +101,27 @@ func convert(filename string) error {
 			kk, kv, err = kvIter.Next()
 			if err == io.EOF {
 				if bk != nil {
+					revertFiles(filename)
 					return errors.New("KV finished iterating before bolt")
 				}
 				break
 			}
 			if err != nil {
+				revertFiles(filename)
 				return err
 			}
 
 			if bk == nil {
+				revertFiles(filename)
 				return errors.New("Bolt finished iterating before kv")
 			}
 
 			if bytes.Compare(kk, bk) != 0 {
+				revertFiles(filename)
 				return fmt.Errorf("Bolt and KV Keys don't match KV: %s, Bolt: %s", kk, bk)
 			}
 			if bytes.Compare(kv, bv) != 0 {
+				revertFiles(filename)
 				return fmt.Errorf("Bolt and KV values don't match KV: %s, Bolt: %s", kv, bv)
 			}
 
@@ -114,5 +130,36 @@ func convert(filename string) error {
 		}
 		return nil
 	})
-	return err
+	if err != nil {
+		revertFiles(filename)
+		return err
+	}
+	return nil
+}
+
+func fileExists(filename string) bool {
+	if _, err := os.Stat(filename); err == nil {
+		return true
+	}
+	return false
+}
+
+func convertedFile(filename string) string {
+	return filepath.Join(filepath.Dir(filename), "."+filepath.Base(filename)+".converted")
+
+}
+
+func revertFiles(filename string) {
+	converted := convertedFile(filename)
+
+	if !fileExists(converted) {
+		return
+	}
+
+	if fileExists(filename) {
+		os.Remove(filename)
+	}
+
+	os.Rename(converted, filename)
+
 }
